Mock user ID rather than send mock users to DB
